+++ test/async/container/channel.rb
@@ -30,6 +30,32 @@ describe Async::Container::Channel do
 		expect(channel.receive).to be_nil
 	end
 	
+	with "#close" do
+		it "can close more than once" do
+			channel.close
+			
+			expect do
+				channel.close
+			end.not.to raise_exception
+		end
+		
+		it "can close the input end after it was already closed" do
+			channel.in.close
+			
+			expect do
+				channel.close_read
+			end.not.to raise_exception
+		end
+		
+		it "can close the output end after it was already closed" do
+			channel.out.close
+			
+			expect do
+				channel.close_write
+			end.not.to raise_exception
+		end
+	end
+	
 	with "timeout" do
 		let(:channel) {subject.new(timeout: 0.001)}